// src/model/DTO/message/query_user_list.dto.ts
import {IsNotEmpty} from 'class-validator';
import {ApiErrorCode} from '../../../config/api-error-code.enum';

export class QueryUserList {
    @IsNotEmpty({ message: 'userId不能为空', context: { errorCode: ApiErrorCode.PARAMS_DELETIONl } })
    userId: any;
    status: number;
    userName: string;
    friends: string;
    groups: string;
    createTime: number;
}